Add a new partition recipe for k8s workers with local storage
Some of the worker nodes in the dse-k8s cluster have some additional SSD
drives that were intended to be used for local storage purposes. They
also have hardware RAID controllers and the existing drives are
configured as RAID1 mirrors.

The hosts' current partman recipe includes the raid1-2dev configuration,
which creates an additional software RAID1 volume, comprised of the two
hardware volumes. This makes the additional SSD drives unusable and is
inefficient.

This commit adds a new partman recipe for kubernetes nodes, which adds
support for hardware RAID devices that present /dev/sda for the O/S and
/dev/sdb for additional storage needs.

It builds on the custom/kubernetes-node-overlay.cfg recipe, but adds
instructions to create an LVM volume group called vg1 from /dev/sdb.

It then configures all of the existing logical volumes to use vg0 from
/dev/sda. Finally, a new logical volume called /srv is created, which
uses vg1.

Bug: T365283
Change-Id: Ib80c345dcc586aae8b8cc3244124d4dc0967a8fb