hdfs: add new an-workers to insetup role
We are transitioning the previous kafka stretch hosts to be part of the hadoop cluster,
The hosts have been renamed to an-worker117[67] and we need to reimage them in preparation
to onboard them to the hdfs cluster. So we shall have them in the insetup::data_engineering role
as we set them up for the analytics_cluster::hadoop::worker role.

Bug: T353788
Change-Id: I98cbd9ce1f8995d6544bed43aaf82546f4781e26