aqs: remove AQSv1 service components
Now that AQS 1 has not been routed to for over a year, remove the AQS
service itself. Don't touch the Cassandra components which are still
in use by the AQS2 services but remove the nodejs-based service and
the various related config.

Note: this removes config for some AQS instances in beta which are no
longer supported.

Bug: T350143
Depends-On: I2564eb44f17d30e7a0310c30984efd6086fccf14
Change-Id: I3fbe8a500b2c8e2dfa0e7a7343e7aac75c435acb