docker_registry_ha: configure the max entries returned for catalog reqs
At the moment when requesting /v2/catalog the HTTP response carries
a Link header that specifies a link carrying "&n=100". It refers to
the default pagination size for a client to use when requesting
the whole set of Docker images, basically what docker-report
follows by default.

The setting is specified here:
  https://github.com/distribution/distribution/blob/
    main/configuration/configuration.go#L209C24-L209C34

In our case the default 100 value is too big, and it causes timeouts
when ATS tries to fetch data from the Docker registry on behalf of
external clients (its timeout is set to 180s). We could bump the ATS
timeout, but 180s is already a lot and it seems wiser to reduce
the pagination to something shorter, to hopefully alleviate the load
on the registry when the entire catalog is retrieved.

This change adds a specific configuration setting for the maxentries
field in the registry's config yaml file.

Change-Id: I8724a2e40ef151668ac4e4d7a731449bc3f2ccfe