gitlab: test defs_from_etcd on the replica
This change tests the defs_from_etcd on one of the GitLab replicas.
Before migrating the hosts to nftables ferm had a BLOCKED_NETS range
which could be used for simple blocking of single IPs.
With nftables this range is gone now so let's sync the ranges
from etcd again.

Bug: T366882
Change-Id: Idf2f25b6afb5e0c5724cc1ca555e5a82c1d013c9