deployment_server: Use the current mw-web image for mwscript-k8s
Previously we read from scap's image build report, and used the
most-recently-built multiversion image. But this has a couple of
drawbacks; it's not necessarily the image being used in prod, and it
only works on the host where the most recent build happened (which is
usually, but not always, the active deployment host).

Instead, check the current state of the mw-web deployment (any
deployment would work; mw-web is chosen semi-arbitrarily) and use that
image. Now mwscript-k8s can be run from any deployment server; the
script itself will run in the RW DC regardless.

Bug: T359127
Change-Id: I5abc09739bd0ab11cd56880012c840f9ac71d0ea