Configure logging correctly for the radosgw service on cephosd servers
We observed that the radosgw service running on cephosd100[1-5] was not
logging to files and was therefore not being rotated by logstash etc.

This commit adds a systemd::syslog resource, which will configure
rsyslog and logrotate correctly.

This class is only applied to the cephosd100[1-5] servers at the moment,
so it does not affect any other clusters.

Hosts: C:ceph::radosgw
Bug: T374477
Change-Id: I6dd9886392ba2b5c922e9401dbeeea29a6c3338e