scap: remove stale production dsh groups
Prunes production mw_web_clusters to only testservers, which should be
the only non-empty group at this point.

The other groups were removed in:
* I89a299f27f4f01725f006c46b0ad39349395cbed
* I26cff0d31abdb399950218f79b8eb28d75036455

Those removals can leave the group file resources in an indeterminate /
stale state [0], so for clarity it's simplest to have scap no longer
consume them.

Discovered following the switchover of the active deployment server to
codfw.

[0] https://phabricator.wikimedia.org/T370962#10180876

Bug: T370962
Change-Id: I6a5960a6521c1d5686c284640873bce9b1f19cf4