ncredir: Add enwp.org/c.enwp.org redirection
enwp.org goes back a long way and has been handled by a lovely community
member that is donating the domain.

Of note, c.enwp.org does not append /wiki like enwp.org does, so we omit
it in the redirect.

Note that this particular domain is a URL shortener and not a typical redirection.

Bug: T332220
Change-Id: If7b588afea240713cc18c2cb3fa87ac1ffdf15ab