swift: avoid rate-limit for the Docker account
The Docker account is currently used by the Docker Registry
to fetch blob data from Swift related to Docker images and
layers.

One of the recurrent jobs that we run is called 'docker-report',
that asks to the Registry the whole catalog/list-of-images to then
decide what image needs to be inspected to generate Debmonitor
reports. The catalog is implemented in the Registry as a paginated
HTTP API, and behind the scenes Swift calls are made to list all
the Docker Images stored. It seems that the Docker Registry issues
close to 30 rps LIST calls to fetch the catalog, together with live traffic,
almost surely ending up in slowdows due to this ratelimit setting in Swift:

container_listing_ratelimit_200 = 30

It is not clear from the Docker Registry logs if it receive HTTP
responses stating a rate limit, but I suspect that the golang
client used automatically implements some backoff workflow, ending
up in higher an higher response latency. This in turn causes calls
to the Catalog to fail (on the Docker Registry side) due to TTFB hitting
values like 300s (from the grafana dashboard it is clear that
the extra latency is all on the Swift side).

Given how critical the Docker Registry is for production, I'd propose
to add it to the accounts that are not subject to the rate limiting.

Bug: T376285
Change-Id: Ia8bc0ab2a4cd317297db49a2c0a1e6a8a72f8368