gerrit: avoid duplicate declaration error on first setup
The "migration" class was specifically created to allow setting up
a new gerrit server before the application is deployed and the
server becomes production.

Due to a race condition there can be a puppet "duplicate declaration"
error between this class and the main gerrit class that both try to
setup the $gerrit_site directory.

Using ensure_resource instead of file{} should avoid this duplicate
declaration.

Bug: T372804
Change-Id: I2c49c7a429afdc3abb63debe78d3db6901693f5c