p:toolforge::proxy: add toolforge api site config
This enables the toolforge api to be accessed from the external world
through api.beta.toolforge.org or api.toolforge.org.

Currently the only authentication allowed is using client
certificates, that are not forwarded by the toolforge proxy, so you
can't really authenticate yet (that will be enabled at a later stage).

This enables us to use non-authenticated endpoints and allows us to
add a webhook endpoint for automated deployment.

Bug: T362066
Signed-off-by: David Caro <dcaro@wikimedia.org>
Change-Id: Ibbddc94abd82cd64bcea4a1cfa16c5596bb1cd12