gerrit: comment out creation of site dir in migration profile
There is a duplicate declaration puppet error between the "migration"
profile and the main gerrit profile.

This was supposed to ensure that the site dir exists already before
gerrit has been deployed for the very first time on a new machine.

Now it causes an issue. I2c49c7a429afdc3ab was a previous attempt
to fix it that did not work.

Bug: T372804
Change-Id: I653ce9a2bbb8c52de7349afe9bc6541bcfac554b