Switch cephosd1001 to use the nftables based firewall
This is a host override to enabled the nftables based firewall on
cephosd1001, as a canary. Once it has been shown to work, we will be
able to remove this host override and enable it for the role.

Since we use envoyproxy on these hosts, we need to update the firewall
definitions to use the new syntax for firewall_src_nets and
firewall_srange.

Hosts: cephosd1001.eqiad.wmnet
Bug: T327259
Change-Id: I1489dbb9d7f44ff5a5464fce6af0e209150dd598