deployment_server: Read Helm secrets in `mwscript-cleanup`
Use the Kubernetes API to iterate over Helm's secrets, rather than
using `helm list` to list all the releases.

`helm list` doesn't offer pagination over a consistent snapshot, so we
get bad data when the list of releases changes while we're trying to
iterate over it.

This is slightly sensitive to a Helm implementation detail, but only we
only use the secrets' metadata (name and modifiedAt label), not any of
the contents, so we expect it to be stable.

Bug: T376795
Change-Id: I4b2d1bd65239bce5235592f68033917ab858e525