Revert "codesearch: add ports for design and discovery"
This reverts commit 9e27826ebe62. These instances were never defined
in the config files, and thus are been broken under the health
page at https://codesearch.wmcloud.org/_health/ for the last two
years.

We've since added `design/*` to Libraries, and to Everywhere.

I don't know which discovery repos are missing/wanted, but there's
only a small handful at https://gerrit.wikimedia.org/g/wikimedia/disco
which if/when there's demand for it, could be added to one of
'operations' or 'services at WMF'.

Change-Id: I6a7db393254c6ca551f24b118f569170baef4dc5