ci: fix git mirror not fetching branches
The mirrored git repositories are updated on a daily basis. Since
7291ed4dd12 added pruning the tags, the branches are no more updated.

I probably removed the branches refspec cause that overrides the tags
refpsec or prevent it from being added but I missed that in a bare
repository the heads are not added since there are no local branches!

Explicitly add the heads and tags refspec to ensure we fetch both, that
would restore updating branches and thus bring up the mirrors up to
date.

Bug: T376981
Change-Id: I4b4564679282c53cf603b66fae3798b9b6649908