growthexperiments.pp: track dangling records for fr+eswiki hourly
To address the ongoing problem with dangling records for
link-recommendations, and in preparation for adding many more such
records in context of our upcoming annual plan work (T362584), we
are looking into clearing outdated link recommendations in a different
hook. See Ibc178572e8b36cf4b5290915357621eabc04af77 for the change of
the GrowthExperiments code.

We want to enable that change in eswiki first and observe its effects
there. See I34308c27b673a541e473bb011067a6179de4f1a8 for the change
toggling the config option for eswiki. frwiki is choosen as the
"control"-wiki that we can use as a reference. See the comments in
T372337 for more context.

This extended logging is intended to be temporary to understand the
effects of this change and will be removed again once the underlying
problem is resolved.

Bug: T372337
Change-Id: I2fd0bd5af96b4a8b8f1beaea46d9e53f596fbc02