Netbox: better logging for scripts import
Currently, when importing a script or report there is no way to know why
it didn't manage to import it. The script will show up in the script
list, but with no way or executing it.

With that patch (tested on netbox-next), the traceback will show up in
/srv/log/netbox/main.log at import time.

From https://github.com/netbox-community/netbox/issues/12766

Change-Id: Ie188fda54a5a8c6bac809c5cf17d5089e5873370