Reduce the sensitivity of the anycast healthcheck for cephosd/radosgw
We have noticed that the anycast route for the rados gateway services on
the cephosd servers prone to flapping, due to occasional failures of the
health checks.

Whilst we plan to investigate this more fully, we want to decrease the
sensitivity of these checks in the meantime.

This commit will require the check to fail three times in a row, before
the route is withdrawn.

Hosts: O:ceph::server
Bug: T376697
Bug: T330153
Change-Id: I90839f9b365e1b1eee7209c44965253ace111fd9