gerrit: remove Hiera keys on host gerrit2003 that are applied by role
Since I218b9f3d9c4dd9b9 now applied the regular gerrit role instead
of the temporary insetup-role on the new machine;
we can remove all the Hiera keys that are applied by it and don't differ
for this machine.

This just excludes setting a service IP and binding it to the NIC,
enabling monitoring and unmasking the service.

All else is standard now after the initial setup process.

Bug: T372804
Change-Id: I232a3501377e17e32ed9aeb86c809f4373452c2f