phabricator: add numeric UID/GID for vcs systemuser
On initial puppet run after a reimage puppet fails
to create the vcs system user.

The reason given is that the additional group vcs does not exist.

Looking at code in systemd:sysuser there is:

if $gid =~ Pattern[/\A\d+\z/] {
group { $username:

This seems to not have been the case in the past.
Should be related to having the globally reserved UID/GID nowadays.

Bug: T377374
Change-Id: Iffa7e255d0beaf9eb63e0d46ce08b67d682f7c03