mariadb: Default pt-heartbeat to STATEMENT-based replication
While allowing for some edge case to define a different binlog
format, default to STATEMENT based replication for pt-heartbeat,
so REPLACE will always be a REPLACE, and not an insert or update
in row format, at least for production hosts.

This is the root cause of the circular replication incident.

Bug: T375144
Change-Id: I5190fe307b08f181af5d69a5ca208f07126a41c2