logstash: force log field to string
Why: image-suggestions sends the log field as an object but most
projects assume this field is a string.  image-suggestions won the
race and now all other log producers are getting forwarded to the DLQ

What: force log field to be a string if not ECS-compliant.  also,
we'll rotate the partition name so that we restore the service by
force-creating a new index

Bug: T377422
Change-Id: I3d8985c0d3ce76f4755ba855897f9724e306feb1