gerrit: make parameter lfs_sync_dest optional
On a test instance, puppet fails with:

lookup() did not find a value for the name 'profile::gerrit::lfs_sync_dest'

This is even the case when the general switch "lfs_replica_sync" is set
to false. (Ib91dd1e2fdcf)

But if lfs syncing is disabled then it makes no sense having to set
a destination for it.

Bug: T363196
Change-Id: Idb2fd605139ad4c91dcc29717864558a5d7cc61a