airflow: make 'secret_key' configurable
Per Airflow docs,
(https://airflow.apache.org/docs/apache-airflow/stable/howto/
set-config.html)
"secret_key" should be the same value on the webserver
and worker. Since we are beginning to use the current airflow
VMs as workers as part of the Kubernetes migration,
add the 'secret_key' as a configurable value within Puppet.

Hosts: an-test-client1002.eqiad.wmnet, an-airflow1005.eqiad.wmnet
Bug: T374948
Change-Id: I29c13f3da65ff22399c5a692a1ac8bdede37e4e4