tables-catalog: Remodel databases for non-default or non-core tables
We have all sorts of complicated cases. For example, echo tables are in
x1 but for some wikis, they are in the wiki's core db. Some tables are
central in wikishared but except on testwiki, some tables only exist in
two wikis, etc. etc.

This model makes sure all are handled.

Bug: T363581
Change-Id: I6a12387ca56ab614bfbbd5e52e7cc6ced49c8c4c