mediabackups: Setup new host for mediabackups backup1012
backup1012 and backup2012 will be new backup storage hosts
for mediabackups infrastructure. For now, install the software
and configuration at the first one, but do not pool it yet.

backup2012 is not yet ready hardware wise, but will receive the
same treatment when it is ready.

Disable temporarily notifications on backup1012 so that it
doesn't alert while it is being setup.

Bug: T376892
Change-Id: I9c5b08299e578eb7dcea6e7bf30cdc100d4448b1