P:dns:auth: alert if a change was submitted but authdns-update was not run
Currently, if a change was submitted to the operations/dns.git repo but
authdns-update was not run, there is no way for us to know that unless
someone tries to run authdns-update again for an unrelated change.

This is not ideal for a variety of reasons:

- It's possible that a change was submitted but authdns-update was not
  run to push it out to all DNS hosts, which means that gdnsd would
  still be returning the outdated data. The person who submitted the
  change but didn't run authdns-update may assume that the change was
  already rolled out.

- It puts the responsibility of figuring out whether the previous change
  should have been merged on the next person that runs authdns-update.

Thus we fix this by alerting if a change was submitted to the
operations/dns.git repo but not actually pushed out via authdns-update.

Hosts: dns6001.wikimedia.org, dns1004.wikimedia.org
Change-Id: Ia0db2053f80e625d33ec7e649d9866cdb79acd9c