gerrit: use systemd::sysuser, reserved UID/GID, new name for daemon user
If on bookworm, use systemd::sysuser to create the daemon user.

Use the UID/GID 925 that has already been reserved in the past
in Ifcb68e36983101937f3.

This also means using the name "gerrit" instead of "gerrit2"
as requested in the ticket.

Doing this only on bookworm means not having to change anything
on existing prod servers and just migrating over once everything
has been upgraded.

Bug: T338470
Change-Id: If94fdced064f6df7b4fa9fdde0fa5b50f8a2f971