alert: Ensure vopsbot database is synced between the alertmanager hosts
Previously, the rsync::quickdatacopy resource was only applied to passive
alertmanager hosts, which led to an incomplete configuration. This approach
missed the necessary logic for synchronizing the vopsbot database across both
active and passive hosts.

This commit adjusts the configuration to ensure that rsync::quickdatacopy is
applied to both active and passive hosts. The quickdatacopy class installs the
rsyncd service on the active host and sets up the script and timer on the
passive hosts to pull the vopsbot database via rsync.

Bug: T375143
Change-Id: Ie3e016f306581555b525f08cf5c410d61e56c0dd