gerrit2003: set SSH disable_nist_kex to true on new bookwork host
The default for this is "true" which means to NOT opt out of disabling
these kex algorithms.

Only for Gerrit there was a special case to opt out of it because
replication broke in the past with:

org.apache.sshd.common.SshException:
KeyExchange signature verification failed for key type=ssh-rsa

due to an older MINA sshd version.

This should not be an issue anymore since we are now on newer
Gerrit and MINA versions.

But to be sure, flip the switch only on the new bookworm server
that is not in production yet.

Bug: T315942
Change-Id: I920abfcf7e5b908f8b7e71031facbc3641e5b3fa