alert: Ensure the db and schema path is synced between alert hosts
Previously, the rsync::quickdatacopy resource was only syncing the schema.sql
file between the alertmanager hosts, which led to an incomplete synchronization
as well as chroot issues. This approach missed the necessary logic for
synchronizing the directory containing the vopsbot database and the schema file
across both active and passive hosts.

This commit adjusts the configuration to ensure that rsync::quickdatacopy
correctly synchronizes the whole /srv/vopsbot directory between the active and
passive hosts.

Bug: T375143
Change-Id: I18709a83a6a2981bc7f5107c7c6f4a4680d90e91