opensearch: Add resource to define cross-cluster settings
OpenSearch comes with cross-cluster search capabilities that need
to be configured with a set of hosts to bootstrap the
cross-cluster connection with. Add a resource that can be
provided the same instance params used for the main opensearch
resource that will configure the cross-cluster seeds.

This was migrated over from the elasticsearch module.

Bug: T380752
Change-Id: I96dd42b728712f4049b0a227c821ce90d945fd66