Allow pty allocation for cumin ssh keys
This allows to allocate a PTY on the target server when using the cumin
and cloud_cumin ssh keys. It's required by the vm_console WMCS cookbook.

For extra safety, I added a default RequestTTY=no in Cumin's ssh_config,
it can be overridden in the cookbooks that need a TTY.

Bug: T379570
Change-Id: I9cb69fcf30024773f9dd4e6987717927a0b2459c