blackbox/http: deployment sites controlled by input parameter instead of ::site
When the resources are computed to be exported,the tag key is set using ::site.

As a result, the site input parameter of the module doesn’t control
where the checks are deployed, and all the rules are only being
deployed to the sites where the generating roles are currently deployed.

Exported resources are imported by prometheus::blackbox::import_checks,
which correctly uses ::site to determine which resources to import.
I believe this is correct because, in this way, each Prometheus
instance in a PoP will only import the resources declared for that
specific PoP.

Bug: T381561
Change-Id: Ica643091b158bad0fe52542ec24ffa5a17b63056