trafficserver: validate production config in tests
Prior to this change, nothing but human review prevents merging a change
to the config that mw-php-migration.lua considers invalid or a change to
the latter that renders the existing config invalid.

While the script fails static to the last-good config in this case, that
doesn't help if it's the first load [0] - which it will be if the script
has changed and trafficserver is reloaded.

In any case, it's staightforward to assert structural validity in tests,
so let's just do it.

Why not also exercise the "real" config-load path? A couple of reasons,
the main one being mock.helper.lua has already overridden global dofile.

[0] In this case, we fall back to the default config embedded in the
script, which may or may not be close to our desired state.

Bug: T377042
Change-Id: Ifd8250f33b6f6711c7564cf7ef74f17e465480a1