haproxy:benthos: produce msg compatible with our schema guidelines
Following indication on
https://wikitech.wikimedia.org/wiki/Event_Platform/Schemas/Guidelines
the produced messages are now compatible with current WMF schema
guidelines and ready for ingestion in our event platform.

The id field could be useful as generic tag for later identification,
set in haproxy.

The grok pattern has been also modified to require type=str and
an extra step to drop messages in case of parsing error has been added.

If the dt field is present in the input, it is reused.
Otherwise it is set to the current ISO8601 UTC time string.
meta.dt is always set to the current ISO8601 UTC time string.
This matches Event Platform guidelines.

Hosts: cp4039.ulsfo.wmnet,cp4048.ulsfo.wmnet
Bug: T329332
Change-Id: If963417a088ec6ab5eda37934a5fbc2011190d06