trafficserver: add dedicated mapping for querybuilder
What:
This change adds a dedicated mapping entry for all miscweb queryservices
for the /querybuilder path. Currently this should be a noop because all
querybuilder endpoints are served from webserver-misc-sites.discovery.wmnet.
(with the expection of the already migrated query-scholarly).

Why:
WMDE asked to separate routing of query-gui and querybuilder to make
sure both services can be migrated independently. querybuilder is in
active development and query-gui should be migrated first.

Bug: T350793
Change-Id: Icfbc6ca5037ec36ee43ee2e5a4c9bca31e867ceb