prometheus: deploy instances from a single configuration
Move to a central configuration for Prometheus instances: this is the
ground work to be able to spread instances across multiple hosts in
codfw/eqiad.

All instances except for ops and k8s-related instances are migrated, and
deployed to the same hosts they are now.

The change is meant to be functionally a no-op, except to change 'ext'
instance retention to bring it in line with the default.

Bug: T371087
Change-Id: I1fae91fdd06649fbb5c02423db7a8340063300c2