Add caps to allow ceph-csi-cephfs to work with the dumps filesystem
We wlready allow the dse-k8s-csi-cephfs ceph client to access the `dpe`
file system, but we have now created a second file system that will
specifically be used for dumps.

This commit adds the necessary capabilities for the csi plugin to be
able to access this file system.

Hosts: O:ceph::server
Bug: T382490
Change-Id: Ic78f4d0e9b2ac706eeda256fcbaa3da09df16435