profile::tlsproxy::envoy: Explicitly configure retries
Why:
profile::tlsproxy::envoy::retries did not actually enable retries
because per

https://www.envoyproxy.io/docs/envoy/v1.23.12/configuration/http/http_filters/router_filter

for retries to be enabled (and default to 1 per the docs), they need to
be configured at either the route config level or the virtual host retry
policy level. We aren't configuring it at the latter and setting the
aforementioned setting to undef means that it isn't set at the former
level either. Effectively no retries were configured.

What:
Switch to always set num_retries and set it to 1, which is the default.
This is what apparently always was the intent.

This is mostly a NOOP these days, as everything that used to use this
Puppet class is on Kubernetes, which uses a different configuration
scheme, save the mwdebug hosts (and downstream users like
deployment-prep)

Hosts: auto
Bug: T380958
Change-Id: Ia00d62e46928ef8994634c3de5a97b13d85b01e4