ci: Install memcached for MediaWiki success cache
We are experimenting with a "success cache" to potentially speed up CI
scenarios where tests are running against the same working directories
multiple times (e.g. in test and then gate-and-submit). We need an easy
way to centrally store SHA256 keys.

The 10MB memcached store should be plenty as we will only be storing the
SHA256 hex digests as keys with no associated data.

Bug: T383243
Change-Id: Idb2353b6bcad5138d381b3a05b731e06fbae92ea