prometheus: reverse proxy for instances belonging to the host' site too
We want all prometheus hosts to be able to reverse proxy requests for
all instances in the host' site too. For example right now even though
prometheus200[78] don't have any instances allocated to them, they
should still be able to forward requests for all instances to
prometheus200[56] when they are pooled in conftool.

Bug: T371087
Change-Id: I0a0e074092f2fda1c9e4db04be279f094091cb6b