gerrit: give it more time to terminate
Today I took 50 seconds to shutdown Gerrit which is most probably due to
time spent compacting the H2 databases. There are 18 databases and the
driver is allowed up to 15 seconds compaction time for each of them.

In some previous restarts, the shutdown took more than the systemd
default of 1m30s which caused it to SIGKILL Gerrit.

Raise the timeout to 3 minutes which would help completing a graceful
shutdown.

Bug: T323754
Change-Id: I39f4d0b8f8d2542c74ec0f9aab0a21a889ccc92d