Fix some wrong descriptions of old dumps
The 2005 dumps listed there include only SQL dumps of the "cur" table,
not XML files (and thus only versions that were current as of the time
the dump was made, which seems important enough to note), and
also means the dump was clearly taken before the MediaWiki 1.5 migration.

The 2006, 2007, and 2010 dumps include both SQL and XML components, rather than only XML.

https://dumps.wikimedia.org/archive/2002/2002-04-18/wikipedia_dump_20020321.sql.gz
is currently in a folder labeled as containing only UseModWiki
dumps, which is wrong.

Bug: T388472
Change-Id: I5c81b80728f9d6b2ba6d4473deec2498880d5819