prometheus: serve apache vhost on localhost too
Requests for this vhost might come in via https from other prometheus
hosts now.

Since all prometheus hosts will be pooled in the loadbalancer, requests
for a prometheus instance might hit an host that doesn't have said
instance. In that case the request is reverse-proxied (via https,
since there might be auth cookies) to hosts that do have the instance.

Bug: T371087
Change-Id: I99a324013572a26ed6efef7dc321d8f1024ecf4d