Remove firewall rule for rsync on archiva
This seems to have been only used by git-fat, which has been phased out.

Initially only drop the firewall rule, which will make the service
inaccesible, but it very easy to revert in case it should still be needed.

I'll make a second patch to actually remove the rsync setup.

Bug: T367315
Change-Id: Ibc7841f6f65167910b6db9625e92d35861e356d9