puppet-microservice: compatibility with Cumin 5+
Why:
Starting with version 5.0.0 Cumin does already optimize the puppetdb
query performing an extract of only the certname and a group by for it.
The microservice is already performing the same operation, leading to a
malformed query.

What:
Apply the extract+group-by only if not already set (with the proper
parameters) by the caller.

Hosts: auto
Change-Id: I34143dffacdba90833ef219932c3da942f027563