gerrit: add gerrit_abusers to block IPs
This change adds tooling to block IPs/IP-ranges on nftables level
on Gerrit. As a intermediate solution the IPs are stored in hiera
but will be moved to confd/requestctl later on.
The single IP in the abuse list is from a recent incident
and puppetizes what's currently configured manually on gerrit.

Bug: T348734
Bug: T351817
Change-Id: If03d531b0f6816d88eb1f10d3b2ac2b394502b37