Add BGP data collection from network devices over GNMI
This patch adds the OpenConfig BGP paths we are interested in to
our gnmic config to collect it from network devices, as well as
some pipeline processing to format it so we can use it from
Prometheus.

Also deleted the Dell stuff as we are not using those on the
network.

Hosts: netflow7001.magru.wmnet,netflow1002.eqiad.wmnet
Bug: T369384
Change-Id: I6892a70bd09731580f3878adc7279f1ce0c8b745