dumps: Configure snapshot1012 with the dumps trait
We wish to use a custom database configuration for running dumps. This
has been added to mediawiki-config, but depends on the `dumps` trait
being present in the file `/etc/wikimedia-servergroup`

This change will add that file to snapshot1012, which will configure it
to use dbstore1008 for its dump.

The default value will be changed to true for ann snapshot servers,
soon.

Hosts: P:dumps::generation::worker::common, snapshot1012.eqiad.wmnet
Bug: T382947
Change-Id: I51a903a632e0ea7de6ce92bcca723fb3adefd6a7