trafficserver: point spiderpig.wikimedia.org to deployment.eqiad.wmnet
Point the new DNS name spiderpig to the backend that represents
the currently active deployment server.

As stated releng wants to run this website on the deployment server
and for $REASONS there is no deployment.discovery.wmnet but it's
always discovery.eqiad.wmnet even when it currently points to codfw.

Would have to ask the serviceops team why.

deployment.eqiad.wmnet is an alias for deploy2002.codfw.wmnet.

Bug: T383946
Change-Id: I2792eb4703a7ce5faf01d741e03db5b24a795101