librenms: Ensure the cache/data directory belongs to librenms
As of now, LibreNMS stores cache files as librenms:librenms with 0644
permissions. This caused an issue where Puppet was constantly changing the owner
and permissions of those files to www-data with 0660 as permissions.

This commit ensures we only manage the `storage/framework/sessions/cache/data/`
directory without modifying the owner and permissions of the files that belong
to it.

Bug: T384440
Change-Id: I5fcb56650d32b05d78b715b3a42532d1cd5d79f8