alertmanager: add missing route for sre-collab-releng receiver
We had an existing custom receiver for the "meta team" that is
collaboration services X releng.

But we did not have a route that would match a team to this
receiver.

The service using this special receiver has been Gerrit and due
to the missing route a recent alert could not be routed.

Then the fallback was used which is "SRE IRC" which translates
to notifications on the #wikimedia-operations channel.

- add the missing route that says:
  -- if it's critical and this special
  team then actually use the existing custom receiver, which will
  then send email to releng, IRC notifications to all 3 channels
  AND create a task for the collab team.
  -- if it's a warning or severy level task, then only create the
     task and only for the collab team, no notification to releng

- add email notification to the collab team in cases where releng
 also already gets email to their list

This is related / in reaction to I750fb7813f9d5763246a but not
exactly the same because it keeps the once agreed upon idea to
inform both teams and all IRC channels.

Also create new receivers for "collab+releng-warning" and
"collab+releng-task" where email and/or IRC notifications
are sent additionally to the releng team.

Change-Id: Ia1cf5ce1dc59331394df79655cdff8ae058335b6