trafficserver: revert cookie-enrolled traffic to 7.4
This patch reverts all cookie-enrolled traffic that _would have_ been
directed to the 8.1-based "next" releases of mw-api-ext and mw-web back
to the 7.4 releases.

Instructions:
1. Merge and puppet-merge this patch.
2. (expedite) Use cumin to invoke run-puppet-agent on A:cp-text.

Reminder: If you have reduced the capacity of the main releases after
moving a significant fraction of traffic to next, you will need to scale
back up to accommodate the returning traffic.

Bug: T383845
Bug: T385395
Change-Id: Idf45573220d28f6f1b981273413d1c3c78c4e562