Update CentralAuth multi-DC rules for SUL3
* Add Special:CentralLogin to exceptions. Like
  Special:CentralAutoLogin, it's used for redirect chains, often
  immediately after account creation/autocreation, so replag
  could be a problem.
* Remove loginwiki from exceptions. Special:CentralLogin and
  Special:CentralAutoLogin have it fully covered, it's not used
  in other ways during authentication.
* Do NOT add auth.wikimedia.org. Other than central login/autologin,
  normal POST rules should suffice there.

For SUL3, this leaves the GET request for returning from the shared
domain to the local domain, which is the return URL used by
AuthManager for remote providers, canonically Special:UserLogin/return.
Unfortunately due to localization its actual URL is unpredictable.
If this turns out to be a problem, we can artificially make it more
unique; for now, just hope it will be fine (not having
Special:CentralLogin in the exemption list until now didn't seem to
cause any problems either).

Bug: T363695
Change-Id: Iae43c65aa4f938753be7779801419e266902f55b