Add the service_proxy to an-launcher1002
At the moment, certain Airflow tasks use the MW Action API and
eventgate-analytics endpoints directly, e.g.

https://meta.wikimedia.org/w/api.php?action=streamconfigs&constraints=canary_events_enabled=1

When we migrate the scheduler to Kubernetesm, these methods will no
longer work, so they will have to use the services_proxy instead.

This change adds the services_proxy to an-launcher1002, so that we can
start to use it ahead of the migration.

Hosts: an-launcher1002.eqiad.wmnet
Bug: T384329
Change-Id: I59b995ee031dff29a275412b34222e14b81bec4f