Prometheus: change gnmi label rewrite from 'target' to 'source'
This changes the existing rewrite from the 'target' to 'instance'
label for gnmi stats, as with the updated export we no longer
get the 'target' (but the 'source' field has the same info).

Also removing the string processing from the gnmic side I added
on the 'source' field (stripping the end of the fqdn), as I see
this is already being done here so no need to duplicate.

Bug: T369384
Change-Id: I6d089652e9cf22391fcb337be3d990f9b9626a82