dumps: Re-enable the enwiki dumps on snapshot1012
We had temporarily disabled the enwiki dump, owing to its tendency to
cause replication lag in the database server that it was using, leading
to knock-on impacts to production services and on-wiki tools.

We have now switched the database that uses to one of the analytics
replicas, so it is now safe to re-enable the dump.

The next full dump will run on the 1st of February. The partial dump
timer is not activated beyond the 25th of the month.

Hosts: snapshot1012.eqiad.wmnet
Bug: T382947
Bug: T368098
Change-Id: Id166a605026bbc5f35c0fd016f58b85ba17c9a2d