gNMIc: Add BGP stats collection for network devices
Second attempt at this after some performance issues with the first
go and lots of testing/changes.  The basic approach here is the
same as previously, we subscribe to what we need as a 'sample' and
have multiple processors defined to convert the static neighbor
attributes (like peer-as) to tags, as well as some of the string-
based return values to numbers so they can be sent to Prometheus.

Tested in codfw and working well, does increase the VM CPU a little
but well within norms.  CPU is spread across all cores nicely.
That's our busiest site in terms of devices and BGP peers, so it
should be ok everywhere.

The key difference from previous is now we have caching and
grouping of data returned from the routers disabled in gnmic, which
is made possible by using event-value-tag-v2, an updated processor
the gnmic devs added to improve performance for our use-case.

Bug: T369384
Change-Id: I689169e3f444f0c6e7734262df4c0a55d0c45b9a