Add FIDO2-based ssh keys for user cmooney
I've 3 keys here which have been generated with my 3 Yubikeys from
my work laptop.

Similar to my existing key the "private key" file in ~/.ssh/ is
protected with a passphrase for these.  So on first use I need to
enter the passphrase as before, after which it is stored in my
ssh-agent.  No change to before in other words.

These are ed25519-sk type keys though, so that file is not an
actual private key, but instead a "key handle" which needs to
be combined with the key material on the Yubikey (which can't
be extracted) to sucessfully authenticate.  More info on task.

I've moved my existing key to buster_ssh_keys for systems that
are running OpenSSH older than 8.3, which doesn't support sk
keys.

Bug: T385229
Change-Id: I458672341b227ff60f0f31fc587c68926b46d5e7