hiera: enable json logging for benthos
The current HAProxy approach with silent-drop-ped requests is to log
on stderr (journal). Using an intermediate variable to build a json
payload and sending logs to a custom ring will allow Benthos to parse
and send these to EventGate.

Condition is defined on both http-request directives but can be
eventually removed if considered a performance issue.

Hosts: cp4039.ulsfo.wmnet,cp1111.eqiad.wmnet
Bug: T383392
Change-Id: I557f0398d6e521366aaaee2bb4d369065b91cc8f